@@ -6,7 +6,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:8.2.0-rc02'
+        classpath 'com.android.tools.build:gradle:8.2.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.9.10"
         classpath "com.diffplug.spotless:spotless-plugin-gradle:6.14.0"
         classpath 'com.novoda:bintray-release:0.8.0'
@@ -70,12 +70,11 @@ dependencies {
     implementation 'org.bouncycastle:bcmail-jdk15on:1.64'
     implementation 'org.gnu.inet:libidn:1.15'
     implementation 'com.google.zxing:core:3.5.0'
-    implementation 'de.measite.minidns:minidns-hla:0.2.4'
+    implementation 'org.minidns:minidns-hla:1.0.4'
     implementation 'me.leolin:ShortcutBadger:1.1.22@aar'
     implementation 'org.whispersystems:signal-protocol-android:2.6.2'
-    implementation 'com.makeramen:roundedimageview:2.3.0'
     implementation 'jetty:javax.servlet:5.1.12'
-    implementation 'com.google.code.gson:gson:2.9.0'
+    implementation 'com.google.code.gson:gson:2.10'
     implementation 'androidx.multidex:multidex:2.0.1'
     implementation 'androidx.legacy:legacy-support-v13:1.0.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
@@ -107,7 +106,7 @@ dependencies {
     implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
     implementation 'com.google.guava:guava:32.1.3-android'
     implementation 'com.github.AppIntro:AppIntro:6.2.0'
-    implementation 'androidx.browser:browser:1.6.0'
+    implementation 'androidx.browser:browser:1.7.0'
     implementation 'com.otaliastudios:transcoder:0.9.1' // 0.10.4 seems to be buggy
     implementation 'com.github.singpolyma:Better-Link-Movement-Method:4df081e1e4'
     implementation project(':libs:AXML')
@@ -129,8 +128,7 @@ dependencies {
     implementation "com.daimajia.swipelayout:library:1.2.0@aar"
     implementation 'com.nineoldandroids:library:2.4.0'
     implementation "androidx.core:core-ktx:1.12.0"
-    implementation "androidx.emoji2:emoji2-emojipicker:1.4.0"
-    implementation "androidx.compose.material3:material3-android:1.2.0-alpha10"
+    implementation "androidx.compose.material3:material3-android:1.2.0-alpha12"
 }
 
 ext {
@@ -152,8 +150,8 @@ android {
         targetSdkVersion 34
 
         // versionNameSuffix " -beta_(2023-10-26)" // " beta_(XXXX-XX-XX)" // activate for beta versions
-        versionCode 148
-        versionName "1.7.7.3"
+        versionCode 153
+        versionName "1.7.8.2"
         //resConfigs "en"
 
         archivesBaseName += "-$versionName"
@@ -190,8 +188,8 @@ android {
 
     compileOptions {
         coreLibraryDesugaringEnabled true
-        sourceCompatibility JavaVersion.VERSION_1_9
-        targetCompatibility JavaVersion.VERSION_1_9
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
     }
 
     flavorDimensions("distribution")
@@ -265,6 +263,7 @@ android {
     }
     buildFeatures {
         buildConfig true
+        viewBinding true
     }
     android.applicationVariants.all { variant ->
         variant.outputs.each { output ->
