@@ -25,6 +25,10 @@ import android.os.PowerManager;
 import android.preference.PreferenceManager;
 import android.util.Log;
 import com.google.common.base.CharMatcher;
+import com.google.gson.stream.JsonWriter;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.nio.charset.StandardCharsets;
 
 import androidx.core.app.NotificationCompat;
 
@@ -117,124 +121,88 @@ public class ExportBackupService extends Service {
         return Arrays.asList(openIntent, amazeIntent, systemFallBack);
     }
 
-    private static void accountExport(final SQLiteDatabase db, final String uuid, final PrintWriter writer) {
-        final StringBuilder builder = new StringBuilder();
-        final Cursor accountCursor = db.query(Account.TABLENAME, null, Account.UUID + "=?", new String[]{uuid}, null, null, null);
+    private static void accountExport(
+            final SQLiteDatabase db, final String uuid, final JsonWriter writer)
+            throws IOException {
+        final Cursor accountCursor =
+                db.query(
+                        Account.TABLENAME,
+                        null,
+                        Account.UUID + "=?",
+                        new String[] {uuid},
+                        null,
+                        null,
+                        null);
         while (accountCursor != null && accountCursor.moveToNext()) {
-            builder.append("INSERT INTO ").append(Account.TABLENAME).append("(");
+            writer.beginObject();
+            writer.name("table");
+            writer.value(Account.TABLENAME);
+            writer.name("values");
+            writer.beginObject();
             for (int i = 0; i < accountCursor.getColumnCount(); ++i) {
-                if (i != 0) {
-                    builder.append(',');
-                }
-                builder.append(accountCursor.getColumnName(i));
-            }
-            builder.append(") VALUES(");
-            for (int i = 0; i < accountCursor.getColumnCount(); ++i) {
-                if (i != 0) {
-                    builder.append(',');
-                }
+                final String name = accountCursor.getColumnName(i);
+                writer.name(name);
                 final String value = accountCursor.getString(i);
                 if (value == null || Account.ROSTERVERSION.equals(accountCursor.getColumnName(i))) {
-                    builder.append("NULL");
-                } else if (Account.OPTIONS.equals(accountCursor.getColumnName(i)) && value.matches("\\d+")) {
+                    writer.nullValue();
+                } else if (Account.OPTIONS.equals(accountCursor.getColumnName(i))
+                        && value.matches("\\d+")) {
                     int intValue = Integer.parseInt(value);
                     intValue |= 1 << Account.OPTION_DISABLED;
-                    builder.append(intValue);
+                    writer.value(intValue);
                 } else {
-                    appendEscapedSQLString(builder, value);
+                    writer.value(value);
                 }
             }
-            builder.append(")");
-            builder.append(';');
-            builder.append('\n');
+            writer.endObject();
+            writer.endObject();
         }
         if (accountCursor != null) {
             accountCursor.close();
         }
-        writer.append(builder.toString());
     }
-    private static void appendEscapedSQLString(final StringBuilder sb, final String sqlString) {
-        DatabaseUtils.appendEscapedSQLString(sb, CharMatcher.is('\u0000').removeFrom(sqlString));
-    }
-    private static void simpleExport(SQLiteDatabase db, String table, String column, String uuid, PrintWriter writer) {
-        final Cursor cursor = db.query(table, null, column + "=?", new String[]{uuid}, null, null, null);
+
+    private static void simpleExport(
+            final SQLiteDatabase db,
+            final String table,
+            final String column,
+            final String uuid,
+            final JsonWriter writer)
+            throws IOException {
+        final Cursor cursor =
+                db.query(table, null, column + "=?", new String[] {uuid}, null, null, null);
         while (cursor != null && cursor.moveToNext()) {
-            writer.write(cursorToString(table, cursor, PAGE_SIZE));
+            writer.beginObject();
+            writer.name("table");
+            writer.value(table);
+            writer.name("values");
+            writer.beginObject();
+            for (int i = 0; i < cursor.getColumnCount(); ++i) {
+                final String name = cursor.getColumnName(i);
+                writer.name(name);
+                final String value = cursor.getString(i);
+                writer.value(value);
+            }
+            writer.endObject();
+            writer.endObject();
         }
         if (cursor != null) {
             cursor.close();
         }
     }
 
-    public static byte[] getKey(final String password, final byte[] salt) throws InvalidKeySpecException {
+    public static byte[] getKey(final String password, final byte[] salt)
+            throws InvalidKeySpecException {
         final SecretKeyFactory factory;
         try {
             factory = SecretKeyFactory.getInstance("PBKDF2WithHmacSHA1");
         } catch (NoSuchAlgorithmException e) {
             throw new IllegalStateException(e);
         }
-        return factory.generateSecret(new PBEKeySpec(password.toCharArray(), salt, 1024, 128)).getEncoded();
+        return factory.generateSecret(new PBEKeySpec(password.toCharArray(), salt, 1024, 128))
+                .getEncoded();
     }
 
-    private static String cursorToString(final String table, final Cursor cursor, final int max) {
-        return cursorToString(table, cursor, max, false);
-    }
-
-    private static String cursorToString(final String table, final Cursor cursor, int max, boolean ignore) {
-        final boolean identities = SQLiteAxolotlStore.IDENTITIES_TABLENAME.equals(table);
-        StringBuilder builder = new StringBuilder();
-        builder.append("INSERT ");
-        if (ignore) {
-            builder.append("OR IGNORE ");
-        }
-        builder.append("INTO ").append(table).append("(");
-        int skipColumn = -1;
-        for (int i = 0; i < cursor.getColumnCount(); ++i) {
-            final String name = cursor.getColumnName(i);
-            if (identities && SQLiteAxolotlStore.TRUSTED.equals(name)) {
-                skipColumn = i;
-                continue;
-            }
-            if (i != 0) {
-                builder.append(',');
-            }
-            builder.append(name);
-        }
-        builder.append(") VALUES");
-        for (int i = 0; i < max; ++i) {
-            if (i != 0) {
-                builder.append(',');
-            }
-            appendValues(cursor, builder, skipColumn);
-            if (i < max - 1 && !cursor.moveToNext()) {
-                break;
-            }
-        }
-        builder.append(';');
-        builder.append('\n');
-        return builder.toString();
-    }
-
-    private static void appendValues(final Cursor cursor, final StringBuilder builder, final int skipColumn) {
-        builder.append("(");
-        for (int i = 0; i < cursor.getColumnCount(); ++i) {
-            if (i == skipColumn) {
-                continue;
-            }
-            if (i != 0) {
-                builder.append(',');
-            }
-            final String value = cursor.getString(i);
-            if (value == null) {
-                builder.append("NULL");
-            } else if (value.matches("[0-9]+")) {
-                builder.append(value);
-            } else {
-                appendEscapedSQLString(builder, value);
-            }
-        }
-        builder.append(")");
-    }
 
     @Override
     public void onCreate() {
@@ -303,7 +271,8 @@ public class ExportBackupService extends Service {
         return START_NOT_STICKY;
     }
 
-    private void messageExport(SQLiteDatabase db, String uuid, PrintWriter writer, Progress progress) {
+    private void messageExport(final SQLiteDatabase db, final String uuid, final JsonWriter writer, final Progress progress)
+                    throws IOException {
         Cursor cursor;
         if (runsTwentySix()) {
             try {
@@ -329,17 +298,25 @@ public class ExportBackupService extends Service {
         int p = 0;
         try {
             while (cursor != null && cursor.moveToNext()) {
-                writer.write(cursorToString(Message.TABLENAME, cursor, PAGE_SIZE, false));
-                if (i + PAGE_SIZE > size) {
-                    i = size;
-                } else {
-                    i += PAGE_SIZE;
+                writer.beginObject();
+                writer.name("table");
+                writer.value(Message.TABLENAME);
+                writer.name("values");
+                writer.beginObject();
+                for (int j = 0; j < cursor.getColumnCount(); ++j) {
+                    final String name = cursor.getColumnName(j);
+                    writer.name(name);
+                    final String value = cursor.getString(j);
+                    writer.value(value);
                 }
+                writer.endObject();
+                writer.endObject();
                 final int percentage = i * 100 / size;
                 if (p < percentage) {
                     p = percentage;
                     notificationManager.notify(EXPORT_BACKUP_NOTIFICATION_ID, progress.build(p));
                 }
+                i++;
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -350,24 +327,32 @@ public class ExportBackupService extends Service {
         }
     }
 
-    private void messageExportmonocles(SQLiteDatabase db, String uuid, PrintWriter writer, Progress progress) {
+    private void messageExportmonocles(final SQLiteDatabase db, final String uuid, final JsonWriter writer, final Progress progress) throws IOException {
         Cursor cursor = db.rawQuery("select mmessages.* from messages join monocles.messages mmessages using (uuid) join conversations on conversations.uuid=messages.conversationUuid where conversations.accountUuid=?", new String[]{uuid});
         int size = cursor != null ? cursor.getCount() : 0;
         Log.d(Config.LOGTAG, "exporting " + size + " monocles messages for account " + uuid);
         int i = 0;
         int p = 0;
         while (cursor != null && cursor.moveToNext()) {
-            writer.write(cursorToString("monocles." + Message.TABLENAME, cursor, PAGE_SIZE, false));
-            if (i + PAGE_SIZE > size) {
-                i = size;
-            } else {
-                i += PAGE_SIZE;
+            writer.beginObject();
+            writer.name("table");
+            writer.value("monocles." + Message.TABLENAME);
+            writer.name("values");
+            writer.beginObject();
+            for (int j = 0; j < cursor.getColumnCount(); ++j) {
+                final String name = cursor.getColumnName(j);
+                writer.name(name);
+                final String value = cursor.getString(j);
+                writer.value(value);
             }
+            writer.endObject();
+            writer.endObject();
             final int percentage = i * 100 / size;
             if (p < percentage) {
                 p = percentage;
                 notificationManager.notify(EXPORT_BACKUP_NOTIFICATION_ID, progress.build(p));
             }
+            i++;
         }
         if (cursor != null) {
             cursor.close();
@@ -377,17 +362,25 @@ public class ExportBackupService extends Service {
         size = cursor != null ? cursor.getCount() : 0;
         Log.d(Config.LOGTAG, "exporting " + size + " WebXDC updates for account " + uuid);
         while (cursor != null && cursor.moveToNext()) {
-            writer.write(cursorToString("monocles.webxdc_updates", cursor, PAGE_SIZE, false));
-            if (i + PAGE_SIZE > size) {
-                i = size;
-            } else {
-                i += PAGE_SIZE;
+            writer.beginObject();
+            writer.name("table");
+            writer.value("monocles.webxdc_updates");
+            writer.name("values");
+            writer.beginObject();
+            for (int j = 0; j < cursor.getColumnCount(); ++j) {
+                final String name = cursor.getColumnName(j);
+                writer.name(name);
+                final String value = cursor.getString(j);
+                writer.value(value);
             }
+            writer.endObject();
+            writer.endObject();
             final int percentage = i * 100 / size;
             if (p < percentage) {
                 p = percentage;
                 notificationManager.notify(EXPORT_BACKUP_NOTIFICATION_ID, progress.build(p));
             }
+            i++;
         }
         if (cursor != null) {
             cursor.close();
@@ -439,19 +432,23 @@ public class ExportBackupService extends Service {
                 IvParameterSpec ivSpec = new IvParameterSpec(IV);
                 cipher.init(Cipher.ENCRYPT_MODE, keySpec, ivSpec);
                 CipherOutputStream cipherOutputStream = new CipherOutputStream(fileOutputStream, cipher);
-                GZIPOutputStream gzipOutputStream = new GZIPOutputStream(cipherOutputStream);
-                PrintWriter writer = new PrintWriter(gzipOutputStream);
-                SQLiteDatabase db = this.mDatabaseBackend.getReadableDatabase();
+                final GZIPOutputStream gzipOutputStream = new GZIPOutputStream(cipherOutputStream);
+                final JsonWriter jsonWriter =
+                        new JsonWriter(
+                                new OutputStreamWriter(gzipOutputStream, StandardCharsets.UTF_8));
+                jsonWriter.beginArray();
+                final SQLiteDatabase db = this.mDatabaseBackend.getReadableDatabase();
                 final String uuid = account.getUuid();
-                accountExport(db, uuid, writer);
-                simpleExport(db, Conversation.TABLENAME, Conversation.ACCOUNT, uuid, writer);
-                messageExport(db, uuid, writer, progress);
-                if (withmonoclesDb) messageExportmonocles(db, uuid, writer, progress);
+                accountExport(db, uuid, jsonWriter);
+                simpleExport(db, Conversation.TABLENAME, Conversation.ACCOUNT, uuid, jsonWriter);
+                messageExport(db, uuid, jsonWriter, progress);
+                if (withmonoclesDb) messageExportmonocles(db, uuid, jsonWriter, progress);
                 for (String table : Arrays.asList(SQLiteAxolotlStore.PREKEY_TABLENAME, SQLiteAxolotlStore.SIGNED_PREKEY_TABLENAME, SQLiteAxolotlStore.SESSION_TABLENAME, SQLiteAxolotlStore.IDENTITIES_TABLENAME)) {
-                    simpleExport(db, table, SQLiteAxolotlStore.ACCOUNT, uuid, writer);
+                    simpleExport(db, table, SQLiteAxolotlStore.ACCOUNT, uuid, jsonWriter);
                 }
-                writer.flush();
-                writer.close();
+                jsonWriter.endArray();
+                jsonWriter.flush();
+                jsonWriter.close();
                 mediaScannerScanFile(file);
                 Log.d(Config.LOGTAG, "written backup to " + file.getAbsoluteFile());
             } catch (Exception e) {
