@@ -98,7 +98,7 @@ public class SearchActivity extends XmppActivity implements TextWatcher, OnSearc
         }
         super.onCreate(bundle);
         this.binding = DataBindingUtil.setContentView(this, R.layout.activity_search);
-        setSupportActionBar((Toolbar) this.binding.toolbar);
+        setSupportActionBar((Toolbar) this.binding.toolbar.getRoot());
         configureActionBar(getSupportActionBar());
 		this.messageListAdapter = new MessageAdapter(this, this.messages, uuid == null);
         this.messageListAdapter.setOnContactPictureClicked(this);
