@@ -83,7 +83,7 @@ public class TrustKeysActivity extends OmemoActivity implements OnKeyStatusUpdat
         binding.cancelButton.setOnClickListener(mCancelButtonListener);
         binding.saveButton.setOnClickListener(mSaveButtonListener);
 
-        setSupportActionBar((Toolbar) binding.toolbar);
+        setSupportActionBar((Toolbar) binding.toolbar.getRoot());
         configureActionBar(getSupportActionBar());
 
         if (savedInstanceState != null) {
