@@ -47,13 +47,13 @@ public class CallManager {
             ToastCompat.makeText(activity, R.string.disable_tor_to_make_call, ToastCompat.LENGTH_SHORT).show();
             return;
         }
-        if (hasPermissions(REQUEST_START_VIDEO_CALL, activity, Manifest.permission.CAMERA)) {
+        if (hasPermissions(REQUEST_START_VIDEO_CALL, activity, Manifest.permission.CAMERA, Manifest.permission.RECORD_AUDIO)) {
             triggerRtpSession(RtpSessionActivity.ACTION_MAKE_VIDEO_CALL, activity, conversation);
         }
     }
 
     public static void triggerRtpSession(final String action, XmppActivity activity, Conversation conversation) {
-        if (activity.xmppConnectionService.getJingleConnectionManager().isBusy()) {
+        if (activity.xmppConnectionService.getJingleConnectionManager().isBusy() != null) {
             ToastCompat.makeText(activity, R.string.only_one_call_at_a_time, ToastCompat.LENGTH_LONG).show();
             return;
         }
