@@ -1,5 +1,7 @@
 package eu.siacs.conversations.utils;
 
+import androidx.annotation.NonNull;
+
 import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
@@ -8,7 +10,7 @@ import eu.siacs.conversations.xmpp.Jid;
 
 public class BackupFileHeader {
 
-    private static final int VERSION = 1;
+    private static final int VERSION = 2;
 
     private String app;
     private Jid jid;
@@ -17,6 +19,7 @@ public class BackupFileHeader {
     private byte[] salt;
 
 
+    @NonNull
     @Override
     public String toString() {
         return "BackupFileHeader{" +
@@ -47,17 +50,19 @@ public class BackupFileHeader {
 
     public static BackupFileHeader read(DataInputStream inputStream) throws IOException {
         final int version = inputStream.readInt();
-        if (version > VERSION) {
-            throw new IllegalArgumentException("Backup File version was " + version + " but app only supports up to version " + VERSION);
-        }
-        String app = inputStream.readUTF();
-        String jid = inputStream.readUTF();
+        final String app = inputStream.readUTF();
+        final String jid = inputStream.readUTF();
         long timestamp = inputStream.readLong();
-        byte[] iv = new byte[12];
+        final byte[] iv = new byte[12];
         inputStream.readFully(iv);
-        byte[] salt = new byte[16];
+        final byte[] salt = new byte[16];
         inputStream.readFully(salt);
-
+        if (version < VERSION) {
+            throw new OutdatedBackupFileVersion();
+        }
+        if (version != VERSION) {
+            throw new IllegalArgumentException("Backup File version was " + version + " but app only supports version " + VERSION);
+        }
         return new BackupFileHeader(app, Jid.of(jid), timestamp, iv, salt);
 
     }
@@ -80,5 +85,9 @@ public class BackupFileHeader {
 
     public long getTimestamp() {
         return timestamp;
+    }
+
+    public static class OutdatedBackupFileVersion extends RuntimeException {
+
     }
 }
