@@ -28,7 +28,7 @@ public final class Namespace {
     public static final String NICK = "http://jabber.org/protocol/nick";
     public static final String FLEXIBLE_OFFLINE_MESSAGE_RETRIEVAL = "http://jabber.org/protocol/offline";
     public static final String BIND = "urn:ietf:params:xml:ns:xmpp-bind";
-    public static final String BIND2 = "urn:xmpp:bind2:0";
+    public static final String BIND2 = "urn:xmpp:bind:0";
     public static final String STREAM_MANAGEMENT = "urn:xmpp:sm:3";
     public static final String CSI = "urn:xmpp:csi:0";
     public static final String CARBONS = "urn:xmpp:carbons:2";
@@ -68,4 +68,5 @@ public final class Namespace {
     public static final String JINGLE_TRANSPORT_ICE_OPTION = "http://gultsch.de/xmpp/drafts/jingle/transports/ice-udp/option";
     public static final String UNIFIED_PUSH = "http://gultsch.de/xmpp/drafts/unified-push";
     public static final String VCARD4 = "urn:ietf:params:xml:ns:vcard-4.0";
+    public static final String SDP_OFFER_ANSWER = "urn:ietf:rfc:3264";
 }
